Schema migrator should lock schema before migration
This change addresses 3 issues to make schema migration less awful.

* Schema migrator should not run concurrently on multiple threads.
  Migrator acquires an exclusive table lock to avoid it.

* Applying migration scripts should not skip migration scripts that are
  not applied yet but older than the latest applied version. This is
  because a pull-request that has larger version of migration script
  could be merged earlier than another pull-request. Even with this
  case, because obviously there're no dependency between those two
  pull-requests (this is an assumption), all of the migration scripts
  should be applied.

* Migration definition files should be separated into files so that
  commit to DatabaseMigrator.java file doesn't cause many conflicts
  that are not obvious to merge.
