+++ digdag-core/src/main/java/io/digdag/core/database/migrate/Migration_20160623123456_AddUserInfoColumnToRevisions.java
@@ -1,0 +1,14 @@
+package io.digdag.core.database.migrate;
+
+import org.skife.jdbi.v2.Handle;
+
+public class Migration_20160623123456_AddUserInfoColumnToRevisions
+        implements Migration
+{
+    @Override
+    public void migrate(Handle handle, MigrationContext context)
+    {
+        handle.update("alter table revisions" +
+                " add column user_info text");
+    }
+}
